+++ modules/components/src/main/java/io/github/palexdev/mfxcomponents/controls/fab/MFXFabBase.java
@@ -24,6 +24,8 @@ import io.github.palexdev.mfxcomponents.controls.base.MFXSkinBase;
 import io.github.palexdev.mfxcomponents.controls.buttons.MFXButton;
 import io.github.palexdev.mfxcomponents.controls.buttons.MFXElevatedButton;
 import io.github.palexdev.mfxcomponents.skins.MFXFabSkin;
+import io.github.palexdev.mfxcomponents.theming.base.WithVariants;
+import io.github.palexdev.mfxcomponents.theming.enums.FABVariants;
 import io.github.palexdev.mfxcore.base.properties.styleable.StyleableBooleanProperty;
 import io.github.palexdev.mfxcore.observables.When;
 import io.github.palexdev.mfxcore.utils.fx.StyleUtils;
@@ -109,7 +111,14 @@ public class MFXFabBase extends MFXElevatedButton {
 	 * <p>
 	 * Also note that this uses {@link #getFabBehavior()} to get the behavior, since it's 'null' safe and exception safe.
 	 */
-	protected final void extend() {
+	protected void extend() {
+		boolean extended = isExtended();
+		if (extended) {
+			WithVariants.addVariants(this, FABVariants.EXTENDED);
+		} else {
+			WithVariants.removeVariants(this, FABVariants.EXTENDED);
+		}
+
 		Skin<?> skin = getSkin();
 		if (skin == null) {
 			// Let's ensure that no other listeners have been added before...