+++ modules/components/src/main/java/io/github/palexdev/mfxcomponents/theming/enums/FABVariants.java
@@ -34,9 +34,6 @@ public enum FABVariants implements Variant {
 	SURFACE("surface"),
 	SECONDARY("secondary"),
 	TERTIARY("tertiary"),
-	// TODO this is a good way of managing it without relying on PseudoClasses
-	// TODO the only improvement I'd make is to have a bunch of methods here responsible for applying the variant
-	// TODO 'fab-extended' should be applied only after 'fab' is removed
 	EXTENDED("fab-extended"),
 	;